sc-meta cli repro build args refactor
