+++ framework/meta/src/cli.rs
@@ -1,5 +1,6 @@
-mod cli_args_standalone;
+pub mod cli_args_reproducible_builds;
 pub mod cli_args_sender;
+mod cli_args_standalone;
 pub mod cli_args_tx;
 mod cli_args_validate;
 mod cli_standalone_main;
